feat(DEV-1552): refact components
